+++ src/utils/index.ts
@@ -1,7 +1,8 @@
 import pkg from '../../package.json';
 
+export * from './cache';
 export * from './clone';
-export * from './config-loader';
+export * from './config';
 export * from './detectors';
 export * from './paths';
 export * from './types';